+++ Source/ColonyManagerRedux/Helpers/MayRequire.cs
@@ -1,0 +1,8 @@
+namespace ColonyManagerRedux;
+
+public sealed class MayRequireSurvivalistsAdditionsAttribute : MayRequireAttribute
+{
+    public MayRequireSurvivalistsAdditionsAttribute() : base(Constants.SurvivalistsAdditionsModId)
+    {
+    }
+}